virtio: Rename VM virtio component to virtioarm
This avoids conflicting with the new util_libs/virtio library.
