+++ CMakeLists.txt
@@ -31,6 +31,7 @@ add_subdirectory(components/Firewall)
 add_subdirectory(components/StringReverse)
 add_subdirectory(components/UDPServer)
 add_subdirectory(components/Sataserver)
+add_subdirectory(libs/libvirtio camkes-arm-vm/libs/libvirtio)
 
 # Declare connectors with templates.
 # We can rely on the naming scheme being consistent to allow for easier declaration.
