+++ libs/libvirtio/include/virtioarm/virtio_console.h
